* 'printf-integers' of https://github.com/t-8ch/util-linux:
  fdisk: use the correct format specifiers for integer types
  lib/procfs: return uint64_t from procfs_process_get_stat_nth()
  chrt: pass correct integer types to printf
  hardlink: avoid format string error for dev_t
  strutils: fix printf formats

+++ disk-utils/fdisk-list.c
@@ -70,7 +70,7 @@ void list_disk_geometry(struct fdisk_context *cxt)
 	color_scheme_enable("header", UL_COLOR_BOLD);
 	fdisk_info(cxt,	_("Disk %s: %s, %ju bytes, %ju sectors"),
 			fdisk_get_devname(cxt), strsz,
-			bytes, (uintmax_t) fdisk_get_nsectors(cxt));
+			(uintmax_t)bytes, (uintmax_t) fdisk_get_nsectors(cxt));
 	color_disable();
 	free(strsz);
